+++ src/GUI/Pridanie/PridanieParcela.java
@@ -26,10 +26,10 @@ public class PridanieParcela extends JPanel
             {
                 String popis = this.input_popis.getText();
 
-                int vlavoDoleX = Integer.parseInt(this.input_vlavoDoleX.getText());
-                int vlavoDoleY = Integer.parseInt(this.input_vlavoDoleY.getText());
-                int vpravoHoreX = Integer.parseInt(this.input_vpravoHoreX.getText());
-                int vpravoHoreY = Integer.parseInt(this.input_vpravoHoreY.getText());
+                double vlavoDoleX = Double.parseDouble(this.input_vlavoDoleX.getText());
+                double vlavoDoleY = Double.parseDouble(this.input_vlavoDoleY.getText());
+                double vpravoHoreX = Double.parseDouble(this.input_vpravoHoreX.getText());
+                double vpravoHoreY = Double.parseDouble(this.input_vpravoHoreY.getText());
 
                 if (prezenter.skusVlozitParcelu(popis, vlavoDoleX, vlavoDoleY, vpravoHoreX, vpravoHoreY))
                 {
